I removed all methods from my Utils class in hikesuggesterCore module and housed methods in their respective classes. I did so a little haphazardly, but it's not my priority right now.
