+++ hikesuggesterCore/src/main/java/models/MountainForecast.java
@@ -1,9 +1,10 @@
 package models;
 
+import java.time.LocalDateTime;
 import java.util.Date;
 
 public class MountainForecast {
-    private final Date date;
+    private final LocalDateTime date;
     private final int temperature;
     private final int windChill;
     private final int windSpeed;
@@ -13,7 +14,7 @@ public class MountainForecast {
     private final int precipProbability;
     private final int precipAmount;
 
-    public MountainForecast(Date date, int temperature, int windChill, int windSpeed, int windDirection, int humidity, int cloudCover, int precipProbability, int precipAmount) {
+    public MountainForecast(LocalDateTime date, int temperature, int windChill, int windSpeed, int windDirection, int humidity, int cloudCover, int precipProbability, int precipAmount) {
         this.date = date;
         this.temperature = temperature;
         this.windChill = windChill;
@@ -25,7 +26,7 @@ public class MountainForecast {
         this.precipAmount = precipAmount;
     }
 
-    public Date getDate() {
+    public LocalDateTime getDate() {
         return date;
     }
